BackEnd/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
backcall==0.2.0
certifi==2021.5.30
colorama==0.4.4
dataclasses==0.8
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-cors-headers==3.10.1
djangorestframework==3.14.0
djongo==1.3.6
flake8==3.9.2
importlib-metadata==4.6.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.2.0
jupyter-core==4.7.1
mccabe==0.6.1
nest-asyncio==1.5.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2 @ file:///C:/ci/psycopg2_1612298619396/work
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pymongo==3.12.3
python-dateutil==2.8.1
pytz==2021.3
pywin32==301
pyzmq==22.1.0
sqlparse==0.2.4
tornado==6.1
traitlets==4.3.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
wcwidth==0.2.5
wincertstore==0.2
zipp==3.5.0
